// backend/src/enums/responses.enum.ts
export enum ResponsesEnum{
    INVALID_MAIL = "E-mail inválido",
    MAIL_REQUIRED = "E-mail é requerido",
   
    INVALID_NAME = "Nome inválido",
    NAME_REQUIRED = "Nome é requerido",
   
    INVALID_PASSWORD = "Senha inválida",
    PASSWORD_REQUIRED = "Senha é requerida",
    
    INVALID_PROFILE = "Perfil inválido",
   
    INVALID_MENU_NAME = "Nome do menu inválido",
    MENU_NAME_REQUIRED = "Nome do menu é requerido",
  
    INVALID_CATEGORY_NAME = "Nome da categoria inválido",
    CATEGORY_NAME_REQUIRED = "Nome da categoria é requerido",
   
    INVALID_DISHE_NAME = "Nome do prato inválido",
    DISHE_NAME_REQUIRED = "Nome do prato é requerido",
   
    INVALID_DISHE_PRICE = "Preço do prato inválido",
    DISHE_PRICE_REQUIRED = "Preço do prato é requerido",
    
    INVALID_DISHE_MODE = "Modo do prato inválido",
    DISHE_MODE_REQUIRED = "Modo do prato é requerido",
   
    DISHE_DESCRIPTION_REQUIRED = "Descrição do prato é requerida",
   
    DISHE_CATEGORY_REQUIRED = "Categoria do prato é requerida",
    
    INVALID_IMAGE = "Imagem inválida",
    NAME_IS_LONG = "nome muito longo use no máximo 30 caracteres",
    NAME_IS_SHORT = "nome muito curto use no mínimo 4 caracteres",
    PASSWORD_IS_SHORT = "senha muito curta use no minímo 4 digitos",
    PASSWORD_IS_LONG = "senha muito longo use no máximo 20 digitos",
    RESTAURANT_ALREADY_EXISTS = "restaurante já existe",
    REGISTERED_RESTAURANT = "restaurante registrado com sucesso",
    RESTAURANT_NOT_FOUND = "restaurante não encontrado",
    INCORRECT_PASSWORD = "senha incorreta",
    LOGIN_SUCCESSFUL = "logado com sucesso!",
    RESUME_IS_LONG = "descrição muito longa use no máximo 300 caracteres",
    RESUME_IS_SHORT = "descrição muito curta use no mínimo 10 caracteres",
    VERIFY_CODE_TO_NEW_EMAIL = "foi enviado um código de verificação para seu novo e-mail.",
    UPDATED_INFO = "informações atualizadas",
    PROFILE_UPDATED = "perfil atualizado",
    VERIFY_CODE_TO_EMAIL = "foi enviado um código de verificação para seu e-mail",
    NOT_HAVE_CHANGES = "nenhuma alteração ocorreu",
    INVALID_CODE = "código inválido",
    DELETED_RESTAURANT = "restaurante deletado",
    ACCESS_DENIED = "acesso não autorizado",
    MENU_NAME_ALREADY_EXISTS = "cardápio já existe",
    MENU_CREATED = "cardápio criado",
    MENU_NOT_FOUND = "cardápio não encontrado",
    MENU_DELETED = "cardápio deletado",
    MENU_UPDATED = "cardápio atualizado",
    MANY_BOOLEAN_IS_REQUIRED = "booleano 'many' é requerido",
    INVALID_ID = "id inválido",
    INVALID_BODY_OF_REQUEST = "corpo da requisição é inválido",
    EMAIL_OR_NAME_ALREADY_EXISTS = "e-mail ou nome já estão em uso",
    CATEGORY_ALREADY_EXISTS_IN_RESTAURANT = "categoria já existe nesse restaurante",
    CATEGORY_CREATED = "categoria criada",
    CATEGORY_NOT_FOUND = "categoria não encontrada",
    CATEGORY_DELETED = "categoria deletada",
    ID_REQUIRED = "id é requerido",
    DISHE_NOT_FOUND = "prato não encontrado",
    DISHE_DELETED = "prato deletado",
    DISHE_IMAGE_UPLOADED = "imagem do prato atualizada",
    INVALID_CATEGORIES = "categoria inválida",
    DISHE_UPDATED = "prato atualizado",
    DISHE_ALREADY_EXISTS = "prato já existe",
    DISHE_CREATED = "prato criado",
    MENU_ID_IS_REQUIRED = "id do menu é requerido",
    VERY_LOW_PRICE = "preço minimo é 1 (reais)",
    CATEGORIES_REQUIRED = "categoria é requerida",
    INVALID_HEX_COLOR = "INVALID_HEX_COLOR"

}